+++ Coeus/AMSGradRule.cpp
@@ -2,8 +2,8 @@
 
 using namespace Coeus;
 
-AMSGradRule::AMSGradRule(NetworkGradient* p_network_gradient, const double p_alpha, const double p_beta1, const double p_beta2,
-                         const double p_epsilon): IUpdateRule(p_network_gradient, p_alpha), _beta1(p_beta1), _beta2(p_beta2), _epsilon(p_epsilon)
+AMSGradRule::AMSGradRule(NetworkGradient* p_network_gradient, const float p_alpha, const float p_beta1, const float p_beta2,
+                         const float p_epsilon): IUpdateRule(p_network_gradient, p_alpha), _beta1(p_beta1), _beta2(p_beta2), _epsilon(p_epsilon)
 {
 	_m = p_network_gradient->get_empty_params();
 	_v = p_network_gradient->get_empty_params();
@@ -40,7 +40,7 @@ void AMSGradRule::update_momentum(const string& p_id, Tensor& p_gradient) {
 
 }
 
-void AMSGradRule::calc_update(map<string, Tensor>* p_gradient, const double p_alpha) {
+void AMSGradRule::calc_update(map<string, Tensor>* p_gradient, const float p_alpha) {
 	IUpdateRule::calc_update(p_gradient, p_alpha);
 
 	for (auto it = p_gradient->begin(); it != p_gradient->end(); ++it) {
