+++ Coeus/ActivationFunctions.h
@@ -5,15 +5,15 @@ namespace Coeus
 	class __declspec(dllexport) ActivationFunctions
 	{
 	public:
-		static double linear(double p_x);
-		static double binary(double p_x);
-		static double sigmoid(double p_x);
-		static double tanh(double p_x);
-		static double exponential(double p_x);
-		static double softplus(double p_x);
-		static double relu(double p_x);
-		static double kexponential(double p_x);
-		static double gauss(double p_x);
+		static float linear(float p_x);
+		static float binary(float p_x);
+		static float sigmoid(float p_x);
+		static float tanh(float p_x);
+		static float exponential(float p_x);
+		static float softplus(float p_x);
+		static float relu(float p_x);
+		static float kexponential(float p_x);
+		static float gauss(float p_x);
 	};
 }
