+++ Coeus/AdadeltaRule.h
@@ -5,19 +5,19 @@ namespace Coeus {
 	class __declspec(dllexport) AdadeltaRule : public IUpdateRule
 	{
 	public:
-		AdadeltaRule(NetworkGradient* p_network_gradient, double p_alpha, double p_decay = 0.9, double p_epsilon = 1e-8);
+		AdadeltaRule(NetworkGradient* p_network_gradient, float p_alpha, float p_decay = 0.9, float p_epsilon = 1e-8);
 		~AdadeltaRule();
 		IUpdateRule* clone(NetworkGradient* p_network_gradient) override;
 
-		void calc_update(map<string, Tensor>* p_gradient, double p_alpha = 0) override;
+		void calc_update(map<string, Tensor>* p_gradient, float p_alpha = 0) override;
 		void reset() override;
 
 	private:
 		void update_cache(const string& p_id, Tensor &p_gradient);
 		void update_cache_delta(const string& p_id, Tensor &p_gradient);
 
-		double _decay;
-		double _epsilon;
+		float _decay;
+		float _epsilon;
 
 		map<string, Tensor> _cache;
 		map<string, Tensor> _cache_delta;
