+++ Coeus/Adagrad.cpp
@@ -11,6 +11,6 @@ Adagrad::Adagrad(NeuralNetwork* p_network) : GradientAlgorithm(p_network)
 Adagrad::~Adagrad()
 = default;
 
-void Adagrad::init(ICostFunction* p_cost_function, const double p_alpha, const double p_epsilon) {
+void Adagrad::init(ICostFunction* p_cost_function, const float p_alpha, const float p_epsilon) {
 	GradientAlgorithm::init(p_cost_function, new AdagradRule(_network_gradient, p_alpha, p_epsilon));
 }
