+++ Coeus/AdagradRule.cpp
@@ -2,7 +2,7 @@
 
 using namespace Coeus;
 
-AdagradRule::AdagradRule(NetworkGradient* p_network_gradient, const double p_alpha, const double p_epsilon): IUpdateRule(p_network_gradient, p_alpha), _epsilon(p_epsilon)
+AdagradRule::AdagradRule(NetworkGradient* p_network_gradient, const float p_alpha, const float p_epsilon): IUpdateRule(p_network_gradient, p_alpha), _epsilon(p_epsilon)
 {
 	_G = p_network_gradient->get_empty_params();
 }
@@ -10,7 +10,7 @@ AdagradRule::AdagradRule(NetworkGradient* p_network_gradient, const double p_alp
 AdagradRule::~AdagradRule()
 = default;
 
-void AdagradRule::calc_update(map<string, Tensor>* p_gradient, const double p_alpha ) {
+void AdagradRule::calc_update(map<string, Tensor>* p_gradient, const float p_alpha ) {
 	IUpdateRule::calc_update(p_gradient, p_alpha);
 	for (auto it = p_gradient->begin(); it != p_gradient->end(); ++it) {
