+++ Coeus/BPTT.cpp
@@ -12,7 +12,7 @@ BPTT::~BPTT()
 {
 }
 
-double BPTT::train(vector<Tensor*>* p_input, Tensor* p_target) const
+float BPTT::train(vector<Tensor*>* p_input, Tensor* p_target) const
 {
 	//return _gradient_algorithm->train(p_input, p_target);
 	return 0;
