+++ Coeus/BackProph.h
@@ -10,6 +10,6 @@ namespace Coeus {
 		explicit BackProp(NeuralNetwork* p_network);
 		~BackProp();
 
-		void init(ICostFunction* p_cost_function, double p_alpha, double p_momentum = 0, bool p_nesterov = false);
+		void init(ICostFunction* p_cost_function, float p_alpha, float p_momentum = 0, bool p_nesterov = false);
 	};
 }
