+++ Coeus/Base_SOM_learning.h
@@ -25,7 +25,7 @@ namespace Coeus
 		void set_mutex(critical_section* p_mutex) { _mutex = p_mutex; }
 
 	protected:
-		double calc_neighborhood(double p_d, NEIGHBORHOOD_TYPE p_type) const;
+		float calc_neighborhood(float p_d, NEIGHBORHOOD_TYPE p_type) const;
 
 		Base_SOM_params* _params;
 		SOM_analyzer* _som_analyzer;
