+++ Coeus/Base_SOM_params.cpp
@@ -13,7 +13,7 @@ Base_SOM_params::~Base_SOM_params()
 {
 }
 
-void Base_SOM_params::init(const double p_epochs)
+void Base_SOM_params::init(const float p_epochs)
 {
 	_iteration = 0;
 	_epochs = p_epochs;
