+++ Coeus/BinaryActivation.cpp
@@ -12,9 +12,9 @@ BinaryActivation::~BinaryActivation()
 }
 
 Tensor BinaryActivation::activate(Tensor& p_input) {
-	double* arr = Tensor::alloc_arr(p_input.size());
-	double* y = &arr[0];
-	double* x = &p_input.arr()[0];
+	float* arr = Tensor::alloc_arr(p_input.size());
+	float* y = &arr[0];
+	float* x = &p_input.arr()[0];
 
 	for(int i = 0; i < p_input.size(); i++) {
 		*y++ = *x++ > 0 ? 1 : 0;
@@ -24,9 +24,9 @@ Tensor BinaryActivation::activate(Tensor& p_input) {
 }
 
 Tensor BinaryActivation::derivative(Tensor& p_input) {
-	double* arr = Tensor::alloc_arr(p_input.size());
-	double* y = &arr[0];
-	double* x = &p_input.arr()[0];
+	float* arr = Tensor::alloc_arr(p_input.size());
+	float* y = &arr[0];
+	float* x = &p_input.arr()[0];
 
 	for (int i = 0; i < p_input.size(); i++) {
 		*y++ = *x++ > 0 ? 1 : 0;
@@ -35,7 +35,7 @@ Tensor BinaryActivation::derivative(Tensor& p_input) {
 	return Tensor({ p_input.size() }, arr);
 }
 
-double BinaryActivation::activate(const double p_value)
+float BinaryActivation::activate(const float p_value)
 {
 	return p_value > 0 ? 1 : 0;
 }
