+++ Coeus/CountModule.h
@@ -11,7 +11,7 @@ namespace Coeus
 		CountModule(int p_state_space_size);
 		~CountModule();
 
-		double get_reward(Tensor* p_state) const;
+		float get_reward(Tensor* p_state) const;
 		void update(Tensor* p_state) const;
 
 	private:
